Driver implementation changes + Separate BrowserNames enum file + New drivers + Relative paths for drivers and screenshots + Folders and files structure changes + Listener changes + Gitignore changes
