+++ .gitignore
@@ -9,3 +9,4 @@ target/
 
 #Other
 screenshots/
+.allure/
