+++ src/main/java/com/utils/BrowserNames.java
@@ -1,0 +1,5 @@
+package com.utils;
+
+public enum BrowserNames {
+    CHROME, FIREFOX
+}
