+++ src/test/java/com/automationpracticetest/BaseTest.java
@@ -1,6 +1,7 @@
 package com.automationpracticetest;
 
-import com.automationpractice.WebDriverSingleton;
+import com.utils.BrowserNames;
+import com.utils.WebDriverSingleton;
 import org.openqa.selenium.WebDriver;
 import org.testng.ITestContext;
 import org.testng.annotations.BeforeMethod;
@@ -14,7 +15,7 @@ public class BaseTest {
     @BeforeMethod //need to rework this method and investigate usage of multiple BeforeMethods (BaseTest VS PageTest)
     public void setUp(ITestContext context) {
 //        driver = WebDriverSingleton.getDriver(System.getProperty("browser")); //implementation without using ENUM
-        driver = WebDriverSingleton.getDriver(WebDriverSingleton.BrowserNames.CHROME); //implementation for ENUM
+        driver = WebDriverSingleton.getDriver(BrowserNames.CHROME); //implementation for ENUM
         driver.manage().timeouts().implicitlyWait(6, TimeUnit.SECONDS);
         driver.manage().window().maximize();
         context.setAttribute("WebDriverContext", driver);
