+++ src/test/java/com/automationpracticetest/LogInPageTest.java
@@ -2,10 +2,14 @@ package com.automationpracticetest;
 
 import com.automationpractice.LogInPage;
 import com.automationpractice.MyAccountPage;
-import com.automationpractice.WebDriverSingleton;
+import com.utils.TestListener;
+import com.utils.WebDriverSingleton;
+import com.utils.BrowserNames;
 import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
 import org.testng.annotations.*;
+
+@Listeners(TestListener.class)
 
 public class LogInPageTest extends BaseTest {
 
@@ -15,7 +19,7 @@ public class LogInPageTest extends BaseTest {
     @BeforeMethod
     public void setUp2() {
 //        driver = WebDriverSingleton.getDriver(System.getProperty("browser")); //implementation without using ENUM
-        driver = WebDriverSingleton.getDriver(WebDriverSingleton.BrowserNames.CHROME); //implementation for ENUM
+        driver = WebDriverSingleton.getDriver(BrowserNames.CHROME); //implementation for ENUM
         loginPage = new LogInPage(driver);
         System.out.println("BeforeMethod from LogInPageTest");
     }
