+++ src/test/java/com/automationpracticetest/MainPageTest.java
@@ -2,8 +2,9 @@ package com.automationpracticetest;
 
 import com.automationpractice.LogInPage;
 import com.automationpractice.MainPage;
-import com.automationpractice.TestListener;
-import com.automationpractice.WebDriverSingleton;
+import com.utils.TestListener;
+import com.utils.WebDriverSingleton;
+import com.utils.BrowserNames;
 import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
 import org.testng.annotations.*;
@@ -18,7 +19,7 @@ public class MainPageTest extends BaseTest {
     @BeforeMethod
     public void setUp2() {
 //        driver = WebDriverSingleton.getDriver(System.getProperty("browser")); //implementation without using ENUM
-        driver = WebDriverSingleton.getDriver(WebDriverSingleton.BrowserNames.CHROME); //implementation for ENUM
+        driver = WebDriverSingleton.getDriver(BrowserNames.CHROME); //implementation for ENUM
         mainPage = new MainPage(driver);
         System.out.println("BeforeMethod from MainPageTest");
     }
