+++ src/test/java/com/automationpracticetest/MyAccountPageTest.java
@@ -1,9 +1,13 @@
 package com.automationpracticetest;
 
 import com.automationpractice.MyAccountPage;
-import com.automationpractice.WebDriverSingleton;
+import com.utils.TestListener;
+import com.utils.WebDriverSingleton;
+import com.utils.BrowserNames;
 import org.openqa.selenium.WebDriver;
 import org.testng.annotations.*;
+
+@Listeners(TestListener.class)
 
 public class MyAccountPageTest extends BaseTest {
 
@@ -13,7 +17,7 @@ public class MyAccountPageTest extends BaseTest {
     @BeforeMethod
     public void setUp() {
 //        driver = WebDriverSingleton.getDriver(System.getProperty("browser")); //implementation without using ENUM
-        driver = WebDriverSingleton.getDriver(WebDriverSingleton.BrowserNames.CHROME); //implementation for ENUM
+        driver = WebDriverSingleton.getDriver(BrowserNames.CHROME); //implementation for ENUM
         myAccountPage = new MyAccountPage(driver);
     }
