+++ src/test/resources/automationpractice.xml
@@ -2,7 +2,7 @@
 
 <suite name="Test Suite for testing http://automationpractice.com/ website" verbose="1">
     <listeners>
-        <listener class-name="com.automationpractice.TestListener"/>
+        <listener class-name="com.utils.TestListener"/>
     </listeners>
     <test name="Main Page Tests">
         <classes>
